+++ hearth/CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased Changes
 ------------------
 
 * Issue - Fix query param `to_s` for empty arrays.
+* Feature - [Breaking Change] Add `config` to `Hearth::Context` and refactor `logger` and `interceptors` inside of it.
 
 1.0.0.pre3 (2024-05-01)
 ------------------